implement project page,create component for status
